Fix Wifi part 1
-Adding WPA_SUPPLICANT_VERSION := VER_0_8_X to the
boardconfig causes wifi in libhardware_legacy to be built with
libwpa_client.so as a shared library. This is important
because the hisi wifi implementation uses that too but,
it uses it's own custom version: libwpa_client_hisi.so
In the vendor tree we simply cp libwpa_client_hisi.so to
libwpa_client.so and boom! wifi fixed.

-The rest is just moving the config files to the correct
paths...