rootdir: Start to clean up initialization scripts.
-There is alot of customization in here by Huawei including
serveral functions that seem built into init that we
don't know what they do. However from the names of
the functions it seems that they load cust configuration
and modem checks among other things.

-On stock there are several "cust" related .props and xml configs.
We probably need to work out which ones are needs and load the
data where it needs to be.